Move relevant routes to 'Fingerprints Resource'
