+++ .gitignore
@@ -1,3 +1,4 @@
 .DS_Store
 api/node-echoprint-server/config.js
 node_modules/
+npm-debug.log
